Update version in Cargo.toml for new release
Version number in Cargo.toml was updated to 1.1.0 to prepare for a new release following refactoring of the code. The main changes in this release include better performance and readability through use of async/await in API calling and downloading, improved error handling, and the addition of 'tokio' as a new dependency to support asynchronous features.