Add image category as an argument in CLI
Modified the CLI to allow taking image category as an argument thus allowing users to download images of their desired categories. Accordingly, the API request was updated to fetch data based on the provided category. Default file name in args.rs was changed to "output.png" and the version in Cargo.toml was bumped to 1.2.0 to reflect these changes.