Add category module for category validity checking
Implemented a new feature that checks a category input for validity before running the main part of the CLI tool. Introduced a new module, 'category.rs', with a function that checks if a given category is valid i.e., "neko", "kitsune", or "waifu", other inputs will return an error. Integrated this check within 'main.rs'. Also, increment the patch version in Cargo.toml due to this new added capability.