Update rneko version
The version number of rneko in Cargo.toml has been incremented from 1.2.1 to 1.2.2. This is because there were significant changes made to the README.md file, including a revision of the default output image and the addition of a 'category' option. Updating the version number helps keep track of these changes and signifies the growth and advancement of the rneko project.